first test passes, refactored testing system
