+++ src/test/java/extras/BasicAlgorithm.java
@@ -1,0 +1,7 @@
+package extras;
+
+public class BasicAlgorithm {
+    public String Foo(String input) throws Exception {
+        return "Hello " + input;
+    }
+}
